binary_num.py
This is the way I would like to introduce students to binary numbers (along with the activity).  Let me know what you think.  I thought it would be cool to have a scratch program and possibly an android app that does the same thing just to show them what's available.